backend/agents/langchain_medical_insights.py
"""
LangChain-based Medical Insights Agent (Enhanced PoC)

This is an enhanced Proof-of-Concept replacement for the existing
`MedicalInsightsAgent` that uses LangChain primitives to split text,
generate embeddings and persist a local FAISS vectorstore per record.

Key Features:
- ✅ LangChain RecursiveCharacterTextSplitter for smart chunking
- ✅ OpenAI embeddings (ada-002) via LangChain
- ✅ FAISS vectorstore for fast semantic search
- ✅ Persistent storage of vectorstores
- ✅ Graceful fallback if dependencies missing

Notes:
- This module is optional and will fail gracefully if `langchain` or
  related dependencies aren't installed.
- It intentionally does not replace application-level audit/RBAC.
- FAISS provides 10-100x faster search vs. manual numpy cosine similarity
"""
import os
import uuid
import json
from typing import List, Optional, Dict, Any
from io import BytesIO
from supabase import create_client

from sqlalchemy.orm import Session

try:
    from langchain.text_splitter import RecursiveCharacterTextSplitter
    from langchain_openai import OpenAIEmbeddings
    from langchain.docstore.document import Document
    from langchain_community.vectorstores import FAISS
except Exception:
    # If langchain (or parts) are not installed, we'll keep references optional
    RecursiveCharacterTextSplitter = None
    OpenAIEmbeddings = None
    Document = None
    FAISS = None

try:
    from PyPDF2 import PdfReader
    from PIL import Image
    import pytesseract
except Exception:
    PdfReader = None
    Image = None
    pytesseract = None

from .base_agent import BaseAgent
from models import Record, RecordText, Embedding, RecordStatusEnum


class LangChainMedicalInsightsAgent(BaseAgent):
    """PoC Medical Insights using LangChain primitives"""

    def __init__(self):
        super().__init__("LangChainMedicalInsightsAgent")
        self.supabase_client = self._init_supabase_client()
        self.bucket_name = os.getenv("SUPABASE_BUCKET", "healthcare-records")
        self.vstore_root = os.getenv("LANGCHAIN_VSTORE_DIR", "langchain_vstores")
        os.makedirs(self.vstore_root, exist_ok=True)

        # Embeddings wrapper will be created lazily
        self.openai_api_key = os.getenv("OPENAI_API_KEY")

    def _init_supabase_client(self):
        try:
            url = os.getenv("SUPABASE_URL")
            key = os.getenv("SUPABASE_KEY")
            if not url or not key:
                raise ValueError("SUPABASE_URL or SUPABASE_KEY not set")
            return create_client(url, key)
        except Exception as e:
            self.logger.error(f"Failed to init Supabase client: {e}")
            return None

    def download_from_supabase(self, file_url: str) -> Optional[bytes]:
        try:
            file_path = file_url.split(f"{self.bucket_name}/")[-1]
            response = self.supabase_client.storage.from_(self.bucket_name).download(file_path)
            return response
        except Exception as e:
            self.logger.error(f"Supabase download failed: {e}")
            return None

    def _extract_text(self, record, file_content: bytes) -> List[str]:
        texts = []
        try:
            if record.file_type.value == "pdf" and PdfReader:
                pdf_file = BytesIO(file_content)
                reader = PdfReader(pdf_file)
                for i, page in enumerate(reader.pages):
                    text = page.extract_text() or ""
                    if text.strip():
                        texts.append(f"[Page {i+1}]\n{text}")
            elif record.file_type.value == "image" and Image and pytesseract:
                image = Image.open(BytesIO(file_content))
                text = pytesseract.image_to_string(image)
                if text.strip():
                    texts.append(text)
            else:
                # Fallback: try to decode as utf-8 text
                try:
                    txt = file_content.decode('utf-8')
                    if txt.strip():
                        texts.append(txt)
                except Exception:
                    texts.append("[No extractable text]")
        except Exception as e:
            self.logger.error(f"Text extraction error: {e}")
            texts.append("[Extraction error]")

        return texts

    def _get_text_chunks(self, texts: List[str]) -> List[Document]:
        if RecursiveCharacterTextSplitter is None or Document is None:
            raise RuntimeError("LangChain text splitter or Document is not available")

        splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=200)
        docs: List[Document] = []
        for t in texts:
            chunks = splitter.split_text(t)
            for chunk in chunks:
                docs.append(Document(page_content=chunk))

        return docs

    def _get_embeddings(self):
        if OpenAIEmbeddings is None:
            raise RuntimeError("LangChain OpenAIEmbeddings is not installed")

        return OpenAIEmbeddings(openai_api_key=self.openai_api_key)

    async def process_record(self, db: Session, record_id: uuid.UUID) -> Dict[str, Any]:
        """Process record: download, extract, chunk, embed, persist vectorstore."""
        try:
            record = db.query(Record).filter(Record.id == record_id).first()
            if not record:
                return self.handle_error(Exception("Record not found"), "Record lookup")

            file_content = self.download_from_supabase(record.file_url)
            if not file_content:
                return self.handle_error(Exception("Failed to download file"), "Supabase download")

            texts = self._extract_text(record, file_content)

            # Create RecordText rows (for traceability)
            for idx, t in enumerate(texts):
                record_text = RecordText(
                    id=uuid.uuid4(), record_id=record_id,
                    extracted_text=t, chunk_index=idx
                )
                db.add(record_text)

            db.commit()

            # Create LangChain Documents with metadata
            docs = []
            try:
                docs = self._get_text_chunks(texts)
            except Exception as e:
                self.logger.warning(f"Text splitting unavailable: {e}")
                # fallback: one document per text
                from langchain.docstore.document import Document as _Doc
                docs = [ _Doc(page_content=t, metadata={"record_id": str(record_id)}) for t in texts ]

            # Attach metadata
            for d in docs:
                if not getattr(d, "metadata", None):
                    d.metadata = {}
                d.metadata["record_id"] = str(record_id)

            # Embeddings + vectorstore
            try:
                emb = self._get_embeddings()
                if FAISS is None:
                    raise RuntimeError("FAISS vectorstore is not available")

                vectorstore = FAISS.from_documents(docs, emb)
                dst = os.path.join(self.vstore_root, str(record_id))
                os.makedirs(dst, exist_ok=True)
                vectorstore.save_local(dst)
            except Exception as e:
                self.logger.error(f"Vectorstore creation failed: {e}")
                # proceed without failing the whole operation

            # Update record status
            record.status = RecordStatusEnum.PROCESSED
            db.commit()

            return self.success_response(
                data={"record_id": str(record_id), "texts_extracted": len(texts)},
                message="Record processed with LangChain PoC"
            )

        except Exception as e:
            db.rollback()
            return self.handle_error(e, "LangChain record processing")
