backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
python-dotenv==1.0.1

# Supabase Storage
supabase==2.1.2

# AI & Machine Learning
openai==1.0.0
numpy==2.0.0

# LangChain Stack (v2 agents - RAG, document loading, vectorstores)
langchain==0.1.4
langchain-openai==0.0.8
faiss-cpu==1.7.4.post1
chromadb==0.4.0

# Document Processing (LangChain loaders + legacy support)
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3

# Task Queue (for async processing)
celery==5.3.4
redis==5.0.1

# Optional (uncomment when ready to use)
# twilio==9.0.0
# pgvector==0.3.0
